Fix Delete ObjectStore

+++ demo/js/indexjs.js
@@ -40,7 +40,8 @@
 
            request.onupgradeneeded = function(event) {
               db = event.target.result;
-              db.deleteObjectStore("pumps");
+              var temp = db.transaction("pumps","readwrite").objectStore("pumps");
+              if (temp != null) { db.deleteObjectStore("pumps"); };
               var objectStore = db.createObjectStore("pumps", {keyPath: "ID"});
               var transaction = event.target.transaction
